Clean up, lint
