+++ package/requirements.txt
@@ -1,2 +1,3 @@
 Flask>=1.0, <2.0
 kedro>=0.14.0
+ipython>=7.0.0, <8.0
